my_agent/agent.py
"""
root_agent.py
-------------

Defines the SmartFilterRootAgent that coordinates queries for the smart filtering
system. This agent routes incoming user messages to specialised sub‑agents
based on the category of the request (transport, accommodation or items) and
handles greetings or clarifications. It uses the `transfer_to_agent` directive
understood by the ADK framework to delegate control to sub‑agents when
necessary.
"""

from google.adk.agents import LlmAgent

from .subagents.transport_agent import transport_agent
from .subagents.accommodation_agent import accommodation_agent
from .subagents.item_agent import item_agent

root_agent = LlmAgent(
    model="gemini-2.0-flash",
    name="SmartFilterRootAgent",
    description="Routes listing queries to the correct specialist agent or prompts for rental requirements.",
    instruction=(
        "You are the central coordinator for iShare's smart filtering system.\n"
        "Your mission is to understand each user message, decide whether to respond yourself or delegate, "
        "and ensure the conversation flows smoothly between the user and the specialist agents.\n\n"
        "When a user sends a message, follow these guidelines:\n"
        "1. **Greetings and small talk:** If the user greets you (e.g., 'hi', 'hello', 'hey') or engages in simple small talk, "
        "do not respond with a greeting. Instead, politely prompt the user to enter their rental requirements. For example, "
        "say: 'Please tell me what you are looking for: a vehicle, accommodation or item to rent.' Do not forward these messages to any sub‑agent.\n"
        "2. **Vehicle/transport queries:** If the message is about renting a vehicle or transport (mentions 'car', 'bike', 'van', etc.), "
        "delegate to the TransportAgent by calling `transfer_to_agent` with 'TransportAgent'.\n"
        "3. **Accommodation queries:** If the message is about a place to stay (mentions 'room', 'apartment', 'hotel', 'house', etc.), "
        "delegate to the AccommodationAgent via `transfer_to_agent` with 'AccommodationAgent'.\n"
        "4. **Item queries:** If the message concerns renting a generic item or product (e.g., 'camera', 'laptop', 'tools'), "
        "delegate to the ItemAgent using `transfer_to_agent` with 'ItemAgent'.\n"
        "5. **Ambiguous or unclear requests:** If it isn't clear which category the user needs, ask a clarifying question such as "
        "'Are you looking for a vehicle, accommodation or item to rent?' and wait for their response.\n"
        "6. **Multi‑category or any listing requests:** If the user explicitly asks to search across multiple categories or for 'any' listing, "
        "politely request that they specify which category is most important (vehicle, accommodation or item). "
        "Once they provide a category, delegate as described above. "
        "Do not attempt to combine results yourself; rely on the specialised agents to perform their own searches and, if needed, transfer control further.\n\n"
        "Whenever you ask for clarification and the user responds, read their answer from the conversation context and proceed with the appropriate delegation. "
        "Always delegate domain‑specific queries to the relevant specialist agent and avoid answering them yourself. "
        "If a query falls outside vehicle, accommodation or item rentals, reply with a friendly message explaining that you can help only with those categories."
    ),
    sub_agents=[transport_agent, accommodation_agent, item_agent],
)

